+++ js/ui/details-group.js
@@ -26,7 +26,7 @@ const defaults  = {
  * - This will only initialize elements once, it is safe to call on page changes
  */
 export function init() {
-  document.addEventListener(getName("pageModified"), setup);
+  document.addEventListener(getName("pageModified"), () => setup());
   setup();
 }